+++ src/main.py
@@ -1,56 +1,40 @@
 import flet as ft
-import funcionalidades.cadastros  as cadastros
-
-
-def pilotos():
-    return ft.DataTable(
-            columns=[
-                ft.DataColumn(ft.Text("Nome")),
-                ft.DataColumn(ft.Text("CPF")),
-                ft.DataColumn(ft.Text("Gênero")),
-                ft.DataColumn(ft.Text("Licença")),
-            ],
-            rows=[
-                ft.DataRow(
-                    cells=[
-                        ft.DataCell(ft.Text("João Ribeiro")),
-                        ft.DataCell(ft.Text("456.789.321-85")),
-                        ft.DataCell(ft.Text("Masculino")),
-                        ft.DataCell(ft.Text("Carga")),
-                    ]
-                ),
-                ft.DataRow(
-                    cells=[
-                        ft.DataCell(ft.Text("Clarice Romano")),
-                        ft.DataCell(ft.Text("782.663.751-28")),
-                        ft.DataCell(ft.Text("Feminino")),
-                        ft.DataCell(ft.Text("Passageiro")),
-                    ]
-                )
-            ]
-        )
-
-
+import funcionalidades.cadastros.piloto  as piloto
+import funcionalidades.cadastros.comissarios  as comissarios
+import funcionalidades.cadastros.aviao  as aviao
+import funcionalidades.home  as home
 
 def update_content(index, body_content):
     body_content.controls.clear() 
 
     if index == 0: 
-        body_content.controls.append(cadastros.home())
+        body_content.controls.append(home.home())
+        
     if index == 1: 
+        body_content.controls.append(ft.Text("Pilotos", size=20))
+        body_content.controls.append(piloto.dialog())
+        body_content.controls.append(ft.Divider())
         body_content.controls.append(ft.Text("Cadastrar Pilotos", size=20))
-        body_content.controls.append(cadastros.cadastrar_piloto())
+        body_content.controls.append(piloto.cadastrar_piloto())
+
 
     elif index == 2: 
+        body_content.controls.append(ft.Text("Comissários", size=20))
+        body_content.controls.append(comissarios.dialog())
+        body_content.controls.append(ft.Divider())
         body_content.controls.append(ft.Text("Cadastrar Comissários", size=20))
-        body_content.controls.append(cadastros.cadastrar_comissario())
+        body_content.controls.append(comissarios.cadastrar_comissario())
 
     elif index == 3:  
         body_content.controls.append(ft.Text("Criar Tripulação", size=20))
 
     elif index == 4: 
+        body_content.controls.append(ft.Text("Aviões", size=20))
+        body_content.controls.append(aviao.dialog())
+        body_content.controls.append(ft.Divider())
         body_content.controls.append(ft.Text("Cadastrar Aviões", size=20))
-        body_content.controls.append(cadastros.cadastrar_aviao())
+        body_content.controls.append(aviao.cadastrar_aviao())
+
 
     elif index == 5:  
         body_content.controls.append(ft.Text("Criar Voo", size=20))
@@ -61,14 +45,16 @@ def main(page: ft.Page):
     page.title = "Gerenciador de Voos"
     page.window_width = 1300
     page.window_height = 700
-    page.scroll = "adaptive"
-    page.vertical_alignment = ft.MainAxisAlignment.CENTER
+    page.scroll = "none"  
+    page.vertical_alignment = ft.MainAxisAlignment.START
     page.update()
-    
-    # body_content = ft.Column([ft.Text("Selecione uma opção no menu")], alignment=ft.MainAxisAlignment.START, expand=True)
-    body_content = ft.Column(alignment=ft.MainAxisAlignment.START, expand=True)
-    body_content.controls.append(cadastros.home())
 
+    body_content = ft.Column(
+        alignment=ft.MainAxisAlignment.START,
+        expand=True,
+        scroll="adaptive",  
+    )
+    body_content.controls.append(home.home())
 
     rail = ft.NavigationRail(
         selected_index=0,
@@ -77,7 +63,7 @@ def main(page: ft.Page):
         min_extended_width=400,
         group_alignment=-0.9,
         destinations=[
-             ft.NavigationRailDestination(
+            ft.NavigationRailDestination(
                 icon=ft.icons.HOME_OUTLINED, selected_icon=ft.icons.HOME_ROUNDED, label="Home"
             ),
             ft.NavigationRailDestination(
@@ -99,18 +85,20 @@ def main(page: ft.Page):
         on_change=lambda e: update_content(e.control.selected_index, body_content),
     )
 
-    page.add(
-        ft.Row(
-            [
-                rail,
-                ft.VerticalDivider(width=1),
-                body_content,
-            ],
-            width=1300,
-            height=700,
-        )
+    layout = ft.Row(
+        [
+            rail, 
+            ft.VerticalDivider(width=1),
+            ft.Container(
+                content=body_content,
+                expand=True,  
+            ),
+        ],
+        expand=True,  
     )
+
+    page.add(layout)
     page.update()
 
-ft.app(target=main)
 
+ft.app(target=main)
